api_yamdb/requirements.txt
requests==2.26.0
asgiref==3.6.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.19.0
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
